requirements.txt
# 核心深度学习框架
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0

# 科学计算和统计
scipy>=1.8.0
scikit-learn>=1.1.0

# 数据可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 进度条和日志
tqdm>=4.64.0

# 数据处理
pandas>=1.4.0

# 联邦学习框架（可选）
# flower>=1.0.0  # 如果需要使用Flower框架
# pysyft>=0.6.0  # 如果需要使用PySyft框架

# 自然语言处理（可选，用于真实文本处理）
# transformers>=4.20.0
# tokenizers>=0.12.0

# 实验跟踪（可选）
# wandb>=0.12.0
# tensorboard>=2.9.0

# 开发和测试工具
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# 其他实用工具
pyyaml>=6.0
jsonschema>=4.0.0
click>=8.0.0